# config/banking-api-openapi.yaml
openapi: 3.0.3
info:
  title: Banking Services API
  version: 1.0.0
  description: |
    # Banking Services API
    
    A comprehensive RESTful API for banking operations including account management, 
    transactions, payments, and beneficiary management.
    
    ## Features
    - Account management and balance inquiries
    - Transaction history and search
    - Domestic and international payments
    - Beneficiary management
    - Account statements generation
    
    ## Security
    This API uses OAuth 2.0 for authentication and authorization. All endpoints 
    require a valid access token except for the health check endpoint.
    
    ## Rate Limiting
    Rate limits are enforced based on your subscription tier:
    - Bronze: 10 requests per minute
    - Silver: 100 requests per minute
    - Gold: 1000 requests per minute
    
    ## Support
    For API support, contact: api-support@bankingservices.com
  contact:
    name: API Support Team
    email: api-support@bankingservices.com
    url: https://developer.bankingservices.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.bankingservices.com/terms
servers:
  - url: https://api.bankingservices.com/v1
    description: Production server
  - url: https://sandbox-api.bankingservices.com/v1
    description: Sandbox server for testing

# API Connect v10 Configuration
...
security:
  - oauth2ClientCredentials: []
  - oauth2AuthCode: []
  - apiKey: []
...
paths:
...
  /accounts:
    get:
      tags:
        - Accounts
      summary: List all accounts
      description: Retrieve a list of all accounts for the authenticated customer
      operationId: getAccounts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: accountType
          in: query
          description: Filter by account type
          schema:
            type: string
            enum: [CHECKING, SAVINGS, CREDIT, INVESTMENT]
        - name: status
          in: query
          description: Filter by account status
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, CLOSED, SUSPENDED]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountListResponse'
              example:
                accounts:
                  - accountId: ACC-123456789
                    accountNumber: '****1234'
                    accountType: CHECKING
                    currency: USD
                    status: ACTIVE
                    nickname: Main Checking
                    openedDate: '2020-01-15'
                  - accountId: ACC-987654321
                    accountNumber: '****5678'
                    accountType: SAVINGS
                    currency: USD
                    status: ACTIVE
                    nickname: Emergency Fund
                    openedDate: '2020-01-15'
                pagination:
                  page: 1
                  limit: 10
                  totalPages: 1
                  totalRecords: 2
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /accounts/{accountId}:
    get:
      tags:
        - Accounts
      summary: Get account details
      description: Retrieve detailed information for a specific account
      operationId: getAccountById
      parameters:
        - $ref: '#/components/parameters/AccountIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDetails'
              example:
                accountId: ACC-123456789
                accountNumber: '****1234'
                fullAccountNumber: '1234567890123456'
                accountType: CHECKING
                currency: USD
                status: ACTIVE
                nickname: Main Checking
                openedDate: '2020-01-15'
                branch:
                  branchId: BR-001
                  branchName: Downtown Branch
                  branchCode: DTN001
                availableBalance: 5432.10
                currentBalance: 5432.10
                overdraftLimit: 1000.00
                interestRate: 0.01
                lastTransactionDate: '2026-01-09T10:30:00Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /accounts/{accountId}/balance:
    get:
      tags:
        - Accounts
      summary: Get account balance
      description: Retrieve current balance information for a specific account
      operationId: getAccountBalance
      parameters:
        - $ref: '#/components/parameters/AccountIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
              example:
                accountId: ACC-123456789
                currency: USD
                availableBalance: 5432.10
                currentBalance: 5432.10
                pendingBalance: 0.00
                overdraftLimit: 1000.00
                lastUpdated: '2026-01-09T16:45:00Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /accounts/{accountId}/transactions:
    get:
      tags:
        - Transactions
      summary: Get account transactions
      description: Retrieve transaction history for a specific account
      operationId: getAccountTransactions
      parameters:
        - $ref: '#/components/parameters/AccountIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: fromDate
          in: query
          description: Start date for transaction history (ISO 8601)
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          description: End date for transaction history (ISO 8601)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /accounts/{accountId}/statements:
    get:
      tags:
        - Statements
      summary: List account statements
      description: Retrieve a list of available statements for an account
      operationId: getAccountStatements
      parameters:
        - $ref: '#/components/parameters/AccountIdParam'
        - name: year
          in: query
          description: Filter by year
          schema:
            type: integer
            minimum: 2000
            maximum: 2100
        - name: month
          in: query
          description: Filter by month (1-12)
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatementListResponse'
              example:
                statements:
                  - statementId: STMT-202601-001
                    accountId: ACC-123456789
                    period:
                      startDate: '2026-01-01'
                      endDate: '2026-01-31'
                    generatedDate: '2026-02-01T00:00:00Z'
                    format: PDF
                    size: 245678
                    status: AVAILABLE
                  - statementId: STMT-202512-001
                    accountId: ACC-123456789
                    period:
                      startDate: '2025-12-01'
                      endDate: '2025-12-31'
                    generatedDate: '2026-01-01T00:00:00Z'
                    format: PDF
                    size: 198432
                    status: AVAILABLE
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    oauth2ClientCredentials:
      type: oauth2
      description: OAuth 2.0 Client Credentials flow for server-to-server authentication
      flows:
        clientCredentials:
          tokenUrl: https://auth.bankingservices.com/oauth2/token
          scopes:
            accounts:read: Read account information
            accounts:write: Modify account information
            transactions:read: Read transaction history
            payments:read: Read payment information
            payments:write: Create and manage payments
            beneficiaries:read: Read beneficiary information
            beneficiaries:write: Manage beneficiaries
            statements:read: Access account statements
      x-ibm-configuration:
        provider: oauth-provider
        
    oauth2AuthCode:
      type: oauth2
      description: OAuth 2.0 Authorization Code flow for user authentication
      flows:
        authorizationCode:
          authorizationUrl: https://auth.bankingservices.com/oauth2/authorize
          tokenUrl: https://auth.bankingservices.com/oauth2/token
          scopes:
            accounts:read: Read account information
            accounts:write: Modify account information
            transactions:read: Read transaction history
            payments:read: Read payment information
            payments:write: Create and manage payments
            beneficiaries:read: Read beneficiary information
            beneficiaries:write: Manage beneficiaries
            statements:read: Access account statements
      x-ibm-configuration:
        provider: oauth-provider
        
    apiKey:
      type: apiKey
      description: API key for application identification
...
  parameters:
    AccountIdParam:
      name: accountId
      in: path
      required: true
      description: Unique identifier for the account
      schema:
        type: string
        pattern: '^ACC-[0-9]{9}$'
      example: ACC-123456789
...
    PageParam:
      name: page
      in: query
      description: Page number for pagination (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      example: 10
  schemas:
...
    AccountListResponse:
      type: object
      required:
        - accounts
        - pagination
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
    AccountSummary:
      type: object
      required:
        - accountId
        - accountNumber
        - accountType
        - currency
        - status
      properties:
        accountId:
          type: string
          description: Unique account identifier
          example: ACC-123456789
        accountNumber:
          type: string
          description: Masked account number
          example: '****1234'
        accountType:
          type: string
          enum: [CHECKING, SAVINGS, CREDIT, INVESTMENT]
          description: Type of account
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: ISO 4217 currency code
          example: USD
        status:
          type: string
          enum: [ACTIVE, INACTIVE, CLOSED, SUSPENDED]
          description: Current account status
        nickname:
          type: string
          description: User-defined account nickname
          maxLength: 50
        openedDate:
          type: string
          format: date
          description: Date when account was opened
    AccountDetails:
      allOf:
        - $ref: '#/components/schemas/AccountSummary'
        - type: object
          required:
            - fullAccountNumber
            - availableBalance
            - currentBalance
          properties:
            fullAccountNumber:
              type: string
              description: Full account number (visible only to account owner)
              example: '1234567890123456'
            branch:
              $ref: '#/components/schemas/Branch'
            availableBalance:
              type: number
              format: double
              description: Available balance for transactions
              example: 5432.10
            currentBalance:
              type: number
              format: double
              description: Current account balance
              example: 5432.10
            overdraftLimit:
              type: number
              format: double
              description: Overdraft limit (if applicable)
              example: 1000.00
            interestRate:
              type: number
              format: double
              description: Annual interest rate (as decimal)
              example: 0.01
            lastTransactionDate:
              type: string
              format: date-time
              description: Date and time of last transaction
    Branch:
      type: object
      required:
        - branchId
        - branchName
        - branchCode
      properties:
        branchId:
          type: string
          description: Unique branch identifier
        branchName:
          type: string
          description: Branch name
        branchCode:
          type: string
          description: Branch code
    BalanceResponse:
      type: object
      required:
        - accountId
        - currency
        - availableBalance
        - currentBalance
        - lastUpdated
      properties:
        accountId:
          type: string
          description: Account identifier
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: ISO 4217 currency code
        availableBalance:
          type: number
          format: double
          description: Available balance for transactions
        currentBalance:
          type: number
          format: double
          description: Current account balance
        pendingBalance:
          type: number
          format: double
          description: Pending transactions total
        overdraftLimit:
          type: number
          format: double
          description: Overdraft limit
        lastUpdated:
          type: string
          format: date-time
          description: Last balance update timestamp
    TransactionListResponse:
      type: object
      required:
        - transactions
        - pagination
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
    TransactionSummary:
      type: object
      required:
        - transactionId
        - accountId
        - transactionType
        - amount
        - currency
        - transactionDate
        - status
      properties:
        transactionId:
          type: string
          description: Unique transaction identifier
        accountId:
          type: string
          description: Associated account identifier
        transactionType:
          type: string
          enum: [DEBIT, CREDIT, TRANSFER, PAYMENT, FEE, INTEREST]
          description: Type of transaction
        amount:
          type: number
          format: double
          description: Transaction amount
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: ISO 4217 currency code
        description:
          type: string
          description: Transaction description
        transactionDate:
          type: string
          format: date-time
          description: Transaction date and time
        status:
          type: string
          enum: [COMPLETED, PENDING, FAILED, CANCELLED]
          description: Transaction status
        balance:
          type: number
          format: double
          description: Account balance after transaction
...
    StatementListResponse:
      type: object
      required:
        - statements
      properties:
        statements:
          type: array
          items:
            $ref: '#/components/schemas/StatementSummary'
    StatementSummary:
      type: object
      required:
        - statementId
        - accountId
        - period
        - generatedDate
        - format
        - status
      properties:
        statementId:
          type: string
          description: Unique statement identifier
        accountId:
          type: string
          description: Associated account identifier
        period:
          $ref: '#/components/schemas/Period'
        generatedDate:
          type: string
          format: date-time
          description: Statement generation date
        format:
          type: string
          enum: [PDF, CSV]
          description: Statement format
        size:
          type: integer
          description: File size in bytes
        status:
          type: string
          enum: [AVAILABLE, GENERATING, ARCHIVED]
          description: Statement status
    Period:
      type: object
      required:
        - startDate
        - endDate
      properties:
        startDate:
          type: string
          format: date
          description: Period start date
        endDate:
          type: string
          format: date
          description: Period end date
    Pagination:
      type: object
      required:
        - page
        - limit
        - totalPages
        - totalRecords
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        totalPages:
          type: integer
          description: Total number of pages
        totalRecords:
          type: integer
          description: Total number of records
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              description: Additional error details
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            timestamp:
              type: string
              format: date-time
              description: Error timestamp
            correlationId:
              type: string
              description: Request correlation ID for tracking
  responses:
    BadRequestError:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: Invalid request parameters
              details:
                - field: amount
                  message: Amount must be greater than 0
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired access token
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456
    ForbiddenError:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: Insufficient permissions to access this resource
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456
    NotFoundError:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: The requested resource was not found
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456
...
    RateLimitError:
      description: Too many requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per time window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when the rate limit resets (Unix timestamp)
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Rate limit exceeded. Please try again later.
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_SERVER_ERROR
              message: An unexpected error occurred. Please try again later.
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456

# Made with Bob